# openapi.json
{
  "components": {
    "schemas": {
      "AcceptInvitationRequest": {
        "description": "§6.1 step 3's `[ יש לי קוד הזמנה ]` branch.\n\nThe callback already accepts an `invitation_token`, but that only helps someone who\nhas the code BEFORE they sign in. A parent whose email differs from the invitation by\none character signs in successfully, matches nothing, and needs a way forward that is\nnot \"sign out and start again with a code you were not told to keep\".",
        "properties": {
          "token": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "AcceptInvitationRequest",
        "type": "object"
      },
      "AppAccessOut": {
        "description": "§6.1's two queries, as the client sees them.\n\nTwo booleans and nothing else. A count of students or studios in the *other* app\nwould leak what §6.1 says neither refusal screen may: \"Neither screen leaks whether\nthe account exists in the other app.\"",
        "properties": {
          "parent": {
            "title": "Parent",
            "type": "boolean"
          },
          "staff": {
            "title": "Staff",
            "type": "boolean"
          }
        },
        "required": [
          "staff",
          "parent"
        ],
        "title": "AppAccessOut",
        "type": "object"
      },
      "CallbackRequest": {
        "properties": {
          "code": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "invitation_token": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invitation Token"
          },
          "state": {
            "maxLength": 64,
            "minLength": 1,
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "code",
          "state"
        ],
        "title": "CallbackRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionResponse": {
        "description": "The access token lives in the body, never in a cookie (§10.3).\n\nThe client holds it in memory and replays it, which is what keeps it out of\nautomatic-credential territory -- a cookie-borne access token is sent by the browser\non every request, which is what makes CSRF possible at all. The refresh token is in\nthe Set-Cookie header and never here.",
        "properties": {
          "access": {
            "$ref": "#/components/schemas/AppAccessOut"
          },
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "active_studio_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Studio Id"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "studios": {
            "items": {
              "$ref": "#/components/schemas/StudioMembershipOut"
            },
            "title": "Studios",
            "type": "array"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "access",
          "studios",
          "active_studio_id"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "StudioMembershipOut": {
        "properties": {
          "is_guardian": {
            "title": "Is Guardian",
            "type": "boolean"
          },
          "person_id": {
            "format": "uuid",
            "title": "Person Id",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "studio_id": {
            "format": "uuid",
            "title": "Studio Id",
            "type": "string"
          },
          "studio_is_demo": {
            "title": "Studio Is Demo",
            "type": "boolean"
          },
          "studio_name": {
            "title": "Studio Name",
            "type": "string"
          }
        },
        "required": [
          "studio_id",
          "studio_name",
          "studio_is_demo",
          "person_id",
          "roles",
          "is_guardian"
        ],
        "title": "StudioMembershipOut",
        "type": "object"
      },
...
      "SwitchStudioRequest": {
        "properties": {
          "studio_id": {
            "format": "uuid",
            "title": "Studio Id",
            "type": "string"
          }
        },
        "required": [
          "studio_id"
        ],
        "title": "SwitchStudioRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Studio Manager API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/accept-invitation": {
      "post": {
        "description": "§6.1 step 3 -- '[ יש לי קוד הזמנה ]'.\n\nThe same `accept_invitation` the callback uses, reached by someone who is already\nsigned in. Without it, a correctly-invited parent whose email differs from the\ninvitation by one character has no way forward at all -- and cannot tell their\nsituation apart from a genuine refusal.",
        "operationId": "accept_invitation_code_api_v1_auth_accept_invitation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Invitation Code",
        "tags": [
          "identity"
        ]
      }
    },
...
    "/api/v1/auth/switch-studio": {
      "post": {
        "description": "§5.2 -- 'A person belonging to more than one studio gets a studio switcher.'\n\nThe target is checked against this identity's own memberships. A switch endpoint that\ntrusted its input would be a cross-tenant read with a friendly name.",
        "operationId": "switch_studio_api_v1_auth_switch_studio_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchStudioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Switch Studio",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/{provider}/callback": {
      "post": {
        "description": "§5.2's server-side exchange, and §6.1 step 3's identity resolution.",
        "operationId": "callback_api_v1_auth__provider__callback_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Callback",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/{provider}/start": {
      "get": {
        "description": "§5.2 -- 'a standard top-level redirect, then PKCE code exchange server-side.'\n\nA 307 to the provider and never a rendered interstitial: an interstitial is one step\ncloser to a webview, and Google answers `disallowed_useragent` inside one.",
        "operationId": "start_api_v1_auth__provider__start_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "app",
            "required": false,
            "schema": {
              "default": "parent",
              "title": "App",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "return_path",
            "required": false,
            "schema": {
              "default": "/",
              "title": "Return Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start",
        "tags": [
          "identity"
        ]
      }
    },
...
  }
}